usr/local/bin/app/requirement.txt
# メール送信は標準ライブラリsmtplibで対応可能
# YAMLの読み込みにPyYAMLを使用
PyYAML==6.0
